feat: display building on every user's screen

+++ src/main/java/de/uniks/pioneers/dto/CreateBuildingDto.java
@@ -1,9 +1,9 @@
 package de.uniks.pioneers.dto;
 
 public record CreateBuildingDto(
-        double x,
-        double y,
-        double z,
+        int x,
+        int y,
+        int z,
         int side,
         String type
 ) {
