+++ src/main/java/de/uniks/pioneers/model/Building.java
@@ -1,9 +1,9 @@
 package de.uniks.pioneers.model;
 
 public record Building(
-        double x,
-        double y,
-        double z,
+        int x,
+        int y,
+        int z,
         String _id,
         int side,
         String type,
